+++ UIKitHw/Models/UserSettings.swift
@@ -9,6 +9,7 @@ import Foundation
 
 final class UserSettings {
   
+  // MARK: - Private enum.
   private enum SettingsKeys: String {
     case userName
     case userMail
@@ -16,66 +17,19 @@ final class UserSettings {
     case userPassword
   }
   
-  static var userName: String? {
-    get {
-      return UserDefaults.standard.string(forKey: SettingsKeys.userName.rawValue)
-    }
-    set {
-      let defaults = UserDefaults.standard
-      let key = SettingsKeys.userName.rawValue
-      if let name = newValue {
-        print("value: \(name) was added to key \(key)")
-        defaults.set(newValue, forKey: key)
-      } else {
-        defaults.removeObject(forKey: key)
-      }
-    }
-  }
+  static let defaults = UserDefaults.standard
   
-  static var userMail: String? {
-    get {
-      return UserDefaults.standard.string(forKey: SettingsKeys.userMail.rawValue)
-    }
-    set {
-      let defaults = UserDefaults.standard
-      let key = SettingsKeys.userMail.rawValue
-      if let mail = newValue {
-        print("value: \(mail) was added to key \(key)")
-        defaults.set(newValue, forKey: key)
-      } else {
-        defaults.removeObject(forKey: key)
-      }
-    }
-  }
+  // MARK: - Static properties.
   
-  static var userPhone: String? {
+  static var users: [User] {
     get {
-      return UserDefaults.standard.string(forKey: SettingsKeys.userPhone.rawValue)
+      if let data = try? PropertyListEncoder().encode(self.users) {
+        return UserDefaults.standard.set(data, forKey: "Users")
+          }
     }
     set {
-      let defaults = UserDefaults.standard
-      let key = SettingsKeys.userPhone.rawValue
-      if let phone = newValue {
-        print("value: \(phone) was added to key \(key)")
-        defaults.set(newValue, forKey: key)
-      } else {
-        defaults.removeObject(forKey: key)
-      }
-    }
-  }
-  
-  static var userPassword: String? {
-    get {
-      return UserDefaults.standard.string(forKey: SettingsKeys.userPassword.rawValue)
-    }
-    set {
-      let defaults = UserDefaults.standard
-      let key = SettingsKeys.userPassword.rawValue
-      if let password = newValue {
-        print("value: \(password) was added to key \(key)")
-        defaults.set(newValue, forKey: key)
-      } else {
-        defaults.removeObject(forKey: key)
+      if let data = defaults.data(forKey: "Users") {
+        let array = try! PropertyListDecoder().decode([User].self, from: data)
       }
     }
   }
